Remove appsettings.json with exposed key

+++ Domain/DTOs/Requests/CreateRequestDtos/CreateExaminationRequestDto.cs
@@ -7,7 +7,7 @@ namespace App.Core.DTOs.Requests.CreateRequestDtos
         public required string ExamTitle { get; set; }
         public DateTime ExamDateAndTime { get; set; }
         public short ExamYear { get; set; }
-        public required int Fee { get; set; }
+        public required decimal Fee { get; set; }
         public ExaminationStatus Status { get; set; }
         public ICollection<Guid> Courses { get; set; } = new List<Guid>();
