+++ Domain/DTOs/Requests/UpdateRequestDtos/UpdateExaminationRequestDto.cs
@@ -6,7 +6,7 @@ namespace App.Core.DTOs.Requests.UpdateRequestDtos
     {
         public string? ExamTitle { get; set; }
         public short? ExamYear { get; set; }
-        public int? Fee { get; set; }
+        public decimal? Fee { get; set; }
         public ExaminationStatus? Status { get; set; }
         public DateTime? ExamDateAndTime { get; set; }
         public List<Guid> SelectedCourses { get; set; } = new();
